chore: should format_version and jupytext_version be updated?
I've been using the default that this cookiecutter provides, but I think I'm missing out on features because this hard-coded number is holding me back. It's also copy-pasted from one new Markdown file to the next, so it's hard to manage this way. Can it be a global (per-project) setting?